# loan-onboarding-agentic-demo/services/loan-service/openapi.yaml
openapi: 3.0.0
info:
  title: loan-service API
  description: OpenAPI 3.0 specification auto-generated by MCP Accelerator from Java
    source code (1 controller(s) parsed).
  version: 1.0.0
servers:
- url: http://localhost:8086
  description: Local service instance
tags:
- name: Loan
  description: Operations from LoanController
paths:
  /api/v1/loans/{id}:
    get:
      operationId: get_by_id
      summary: GET /api/v1/loans/{id}
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: integer
    put:
      operationId: update
      summary: PUT /api/v1/loans/{id}
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanApplication'
  /api/v1/loans/customer/{cid}:
    get:
      operationId: by_customer
      summary: GET /api/v1/loans/customer/{cid}
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: cid
        in: path
        required: true
        description: ''
        schema:
          type: integer
  /api/v1/loans/status/{status}:
    get:
      operationId: by_status
      summary: GET /api/v1/loans/status/{status}
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: status
        in: path
        required: true
        description: ''
        schema:
          type: string
  /api/v1/loans/{id}/submit:
    put:
      operationId: submit
      summary: PUT /api/v1/loans/{id}/submit
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: integer
  /api/v1/loans/{id}/review:
    put:
      operationId: review
      summary: PUT /api/v1/loans/{id}/review
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: integer
  /api/v1/loans/{id}/approve:
    put:
      operationId: approve
      summary: PUT /api/v1/loans/{id}/approve
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: integer
  /api/v1/loans/{id}/reject:
    put:
      operationId: reject
      summary: PUT /api/v1/loans/{id}/reject
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: integer
  /api/v1/loans/{id}/disburse:
    put:
      operationId: disburse
      summary: PUT /api/v1/loans/{id}/disburse
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: integer
  /api/v1/loans/{id}/status:
    put:
      operationId: update_status
      summary: PUT /api/v1/loans/{id}/status
      tags:
      - Loan
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Internal server error
      parameters:
      - name: id
        in: path
        required: true
        description: ''
        schema:
          type: integer
components:
  schemas:
    LoanApplication:
      type: object
      description: 'Request body DTO: LoanApplication'